fix(b-human): import both `image` and `lower_image`
by resetting the other in the `InputData` to `None`, when the current
converter does not handle it. This way the available one will be
converted and not only the "upper" image.

The double import of both images results in the images to almost have
the same timestamps and are resampled separately with
`IMAGE_MAX_RESAMPLING_RATE_HZ`, effectively doubling it.